# admin/internal/items/http/app/docs/swagger.yaml
basePath: /
definitions:
  models.Feed:
    properties:
      category_id:
        description: Fixed typo
        type: integer
      description:
        type: string
      link:
        type: string
      logo_url:
        description: Fixed to snake_case
        type: string
      max_items:
        type: integer
      priority:
        type: integer
      title:
        type: string
    type: object
  models.FeedCategory:
    properties:
      icon_url:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  models.FeedItemResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      feed_id:
        type: integer
      id:
        type: integer
      image_url:
        type: string
      published_at:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  models.FeedResponse:
    properties:
      category:
        description: Fixed typo
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      last_refreshed:
        type: string
      link:
        type: string
      logo_url:
        description: Fixed to snake_case
        type: string
      max_items:
        type: integer
      priority:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
  models.UpdateFeed:
    properties:
      category_id:
        description: Fixed typo
        type: integer
      description:
        type: string
      id:
        description: Fixed key to be Id
        type: integer
      link:
        type: string
      logo_url:
        description: Fixed to snake_case
        type: string
      max_items:
        type: integer
      priority:
        type: integer
      title:
        type: string
    type: object
info:
  contact: {}
  description: This is a posts API for Pro-Sphere
  title: Pro-Sphere Posts API
  version: "1.0"
paths:
  /feedTest/{id}:
    get:
      consumes:
      - application/json
      description: Fetcher
      parameters:
      - description: Feed ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.FeedItemResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: This endpoint is used to test the feed fetcher
      tags:
      - Feed Fetcher (test)
  /feedcategories:
    get:
      description: Get all feed categories with pagination
      parameters:
      - description: limit
        in: query
        name: limit
        type: integer
      - description: page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.FeedCategory'
            type: array
        "204":
          description: No Content
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all categories
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: Create a new feed category
      parameters:
      - description: category
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/models.FeedCategory'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.FeedCategory'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new category
      tags:
      - Categories
  /feedcategories/{id}:
    delete:
      description: Delete a feed category by its ID
      parameters:
      - description: category_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a category by ID
      tags:
      - Categories
    get:
      description: Get a feed category by its ID
      parameters:
      - description: category_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.FeedCategory'
        "204":
          description: No Content
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get category by ID
      tags:
      - Categories
    put:
      consumes:
      - application/json
      description: Update a feed category by its ID
      parameters:
      - description: category
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/models.FeedCategory'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.FeedCategory'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update an existing category
      tags:
      - Categories
  /feeds:
    get:
      description: Get all feeds with pagination
      parameters:
      - description: limit
        in: query
        name: limit
        type: integer
      - description: page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.FeedResponse'
            type: array
        "204":
          description: No Content
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all feeds
      tags:
      - Feeds
    post:
      consumes:
      - application/json
      description: Create a new feed
      parameters:
      - description: feed
        in: body
        name: feed
        required: true
        schema:
          $ref: '#/definitions/models.Feed'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.FeedResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new feed
      tags:
      - Feeds
    put:
      consumes:
      - application/json
      description: Update a feed by its ID
      parameters:
      - description: feed
        in: body
        name: feed
        required: true
        schema:
          $ref: '#/definitions/models.UpdateFeed'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.FeedResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update an existing feed
      tags:
      - Feeds
  /feeds/{id}:
    delete:
      description: Delete a feed by its ID
      parameters:
      - description: feed_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a feed by ID
      tags:
      - Feeds
    get:
      description: Get a feed by its ID
      parameters:
      - description: feed_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.FeedResponse'
        "204":
          description: No Content
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get feed by ID
      tags:
      - Feeds
schemes:
- http
swagger: "2.0"